refactor(telemetry): split telemetry.py into telemetry/ package
Split 349-LOC module into focused submodules:
- telemetry/_state.py: shared mutable state (flags, probe timer)
- telemetry/logging.py: structlog config, log processors
- telemetry/tracing.py: init/shutdown, get_tracer
- telemetry/exporters.py: OTLP exporter creation, connectivity probing
- telemetry/__init__.py: re-exports all public symbols

Tests updated to patch new submodule locations for state and
connectivity checks.
