+++ tests/test_telemetry.py
@@ -8,6 +8,7 @@ from opentelemetry import metrics, trace
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.trace import TracerProvider
 
+from gitlab_copilot_agent.telemetry import _state as tel_state
 from gitlab_copilot_agent.telemetry import (
     add_trace_context,
     configure_logging,
@@ -32,7 +33,10 @@ def test_init_telemetry_configures_provider(monkeypatch: pytest.MonkeyPatch) -> 
         patch("opentelemetry.exporter.otlp.proto.grpc.trace_exporter.OTLPSpanExporter"),
         patch("opentelemetry.exporter.otlp.proto.grpc.metric_exporter.OTLPMetricExporter"),
         patch("opentelemetry.exporter.otlp.proto.grpc._log_exporter.OTLPLogExporter"),
-        patch("gitlab_copilot_agent.telemetry._check_connectivity", return_value=True),
+        patch(
+            "gitlab_copilot_agent.telemetry.tracing.check_connectivity",
+            return_value=True,
+        ),
     ):
         init_telemetry()
         assert isinstance(trace.get_tracer_provider(), TracerProvider)
@@ -78,11 +82,9 @@ def test_emit_to_otel_logs_noop_when_unconfigured() -> None:
 
 
 def test_emit_to_otel_logs_emits_when_configured(monkeypatch: pytest.MonkeyPatch) -> None:
-    import gitlab_copilot_agent.telemetry as tel_mod
-
-    monkeypatch.setattr(tel_mod, "_otel_logging_configured", True)
-    monkeypatch.setattr(tel_mod, "_collector_reachable", False)
-    with patch("gitlab_copilot_agent.telemetry.logging") as mock_logging:
+    monkeypatch.setattr(tel_state, "otel_logging_configured", True)
+    monkeypatch.setattr(tel_state, "collector_reachable", False)
+    with patch("gitlab_copilot_agent.telemetry.logging.logging") as mock_logging:
         mock_logger = mock_logging.getLogger.return_value
         mock_logging.INFO = 20
         event_dict: dict[str, object] = {"event": "clone_done", "level": "info", "branch": "main"}
@@ -95,11 +97,9 @@ def test_emit_to_otel_logs_does_not_drop_event_when_collector_reachable(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     """Logs should always pass through to stdout even when OTLP is active."""
-    import gitlab_copilot_agent.telemetry as tel_mod
-
-    monkeypatch.setattr(tel_mod, "_otel_logging_configured", True)
-    monkeypatch.setattr(tel_mod, "_collector_reachable", True)
-    with patch("gitlab_copilot_agent.telemetry.logging") as mock_logging:
+    monkeypatch.setattr(tel_state, "otel_logging_configured", True)
+    monkeypatch.setattr(tel_state, "collector_reachable", True)
+    with patch("gitlab_copilot_agent.telemetry.logging.logging") as mock_logging:
         mock_logger = mock_logging.getLogger.return_value
         mock_logging.INFO = 20
         event_dict: dict[str, object] = {"event": "clone_done", "level": "info", "branch": "main"}
@@ -133,18 +133,19 @@ def test_configure_logging_suppresses_otel_exporters() -> None:
 
 def test_init_telemetry_starts_probe_when_unavailable(monkeypatch: pytest.MonkeyPatch) -> None:
     """When collector is unreachable at init, a background probe timer is scheduled."""
-    import gitlab_copilot_agent.telemetry as tel_mod
-
     monkeypatch.setenv("OTEL_EXPORTER_OTLP_ENDPOINT", "http://localhost:4317")
 
     with (
         patch("opentelemetry.exporter.otlp.proto.grpc.trace_exporter.OTLPSpanExporter"),
         patch("opentelemetry.exporter.otlp.proto.grpc.metric_exporter.OTLPMetricExporter"),
         patch("opentelemetry.exporter.otlp.proto.grpc._log_exporter.OTLPLogExporter"),
-        patch.object(tel_mod, "_check_connectivity", return_value=False),
+        patch(
+            "gitlab_copilot_agent.telemetry.tracing.check_connectivity",
+            return_value=False,
+        ),
     ):
         init_telemetry()
-        assert tel_mod._probe_timer is not None
+        assert tel_state.probe_timer is not None
         shutdown_telemetry()
     # Reset providers for other tests
     from opentelemetry.metrics._internal import _ProxyMeterProvider
@@ -155,19 +156,20 @@ def test_init_telemetry_starts_probe_when_unavailable(monkeypatch: pytest.Monkey
 
 def test_init_telemetry_no_probe_when_available(monkeypatch: pytest.MonkeyPatch) -> None:
     """When collector is reachable at init, no probe timer is started."""
-    import gitlab_copilot_agent.telemetry as tel_mod
-
     monkeypatch.setenv("OTEL_EXPORTER_OTLP_ENDPOINT", "http://localhost:4317")
 
     with (
         patch("opentelemetry.exporter.otlp.proto.grpc.trace_exporter.OTLPSpanExporter"),
         patch("opentelemetry.exporter.otlp.proto.grpc.metric_exporter.OTLPMetricExporter"),
         patch("opentelemetry.exporter.otlp.proto.grpc._log_exporter.OTLPLogExporter"),
-        patch.object(tel_mod, "_check_connectivity", return_value=True),
+        patch(
+            "gitlab_copilot_agent.telemetry.tracing.check_connectivity",
+            return_value=True,
+        ),
     ):
         init_telemetry()
-        assert tel_mod._probe_timer is None
-        assert tel_mod._collector_reachable is True
+        assert tel_state.probe_timer is None
+        assert tel_state.collector_reachable is True
         shutdown_telemetry()
     from opentelemetry.metrics._internal import _ProxyMeterProvider
 
@@ -177,8 +179,6 @@ def test_init_telemetry_no_probe_when_available(monkeypatch: pytest.MonkeyPatch)
 
 def test_init_telemetry_noop_skips_probe(monkeypatch: pytest.MonkeyPatch) -> None:
     """When OTEL endpoint is not set, no probe is scheduled."""
-    import gitlab_copilot_agent.telemetry as tel_mod
-
     monkeypatch.delenv("OTEL_EXPORTER_OTLP_ENDPOINT", raising=False)
     init_telemetry()
-    assert tel_mod._probe_timer is None
+    assert tel_state.probe_timer is None
